set up models and wired endpoints
